extract footer and nav to single components
